requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.5

# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
alembic==1.14.1
greenlet==3.1.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# AI / LLM
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-openai==0.3.0
langchain-community==0.3.14
google-generativeai==0.8.4

# Data Analysis
pandas==2.2.3

# Utilities
python-dotenv==1.0.1
pydantic-settings==2.7.1
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
openpyxl==3.1.5
